requirements.txt
dataclasses==0.6
decorator==5.1.1
frozenlist==1.3.0
huggingface-hub==0.12.0
importlib-metadata==4.10.1
importlib-resources==5.4.0
matplotlib==3.2.1
matplotlib-inline==0.1.3
numpy==1.21.5
pandas==1.3.5
pytorch-crf==0.7.2
scikit-learn==0.22.2.post1
scipy==1.4.1
tokenizers==0.13.2
torch==1.5.1
torchvision==0.6.1
tqdm==4.61.1
traitlets==5.1.1
transformers==4.26.0
